\name{lyne-package}
\alias{lyne}
\docType{package}
\title{
lyne - Lineage Dynamics for Epigenetic Marks
}
\description{
\code{lyne} implements continuous time Markov models 
on cell lineage trees, meant to describe the
changes of epigenetic marks like 5-mC methylation.
}
\details{
\tabular{ll}{
Package: \tab lyne\cr
Type: \tab Package\cr
License: \tab Artistic-2.0\cr
}
}
\author{
Dennis Kostka\cr
Maintainer: Dennis Kostka <kostka@pitt.edu>
}
\references{
JA Capra and D Kostka (2014) 
\emph{Modeling DNA methylation dynamics with approaches from phylogenetics.}
arXiv: \url{http://arxiv.org/abs/1404.3223} 
}
\keyword{ Tree model}
\keyword{ Continuous time Markov chain}
\keyword{ DNA methylation}
